+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/services/AnswerService.java
@@ -2,5 +2,9 @@ package ru.dpopkov.knowthenics.services;
 
 import ru.dpopkov.knowthenics.model.Answer;
 
+import java.util.Set;
+
 public interface AnswerService extends CrudService<Answer, Long> {
+
+    Set<Answer> findAllByWordingEnLike(String searchString);
 }