+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/services/map/AnswerMapService.java
@@ -5,7 +5,17 @@ import org.springframework.stereotype.Service;
 import ru.dpopkov.knowthenics.model.Answer;
 import ru.dpopkov.knowthenics.services.AnswerService;
 
+import java.util.Set;
+import java.util.stream.Collectors;
+
 @Service
 @Profile({"default", "map-service"})
 public class AnswerMapService extends AbstractMapService<Answer> implements AnswerService {
+
+    @Override
+    public Set<Answer> findAllByWordingEnLike(String searchString) {
+        return super.findAll().stream()
+                .filter(a -> a.getWordingEn() != null && a.getWordingEn().contains(searchString))
+                .collect(Collectors.toSet());
+    }
 }