+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/services/springdatajpa/AnswerSDJpaService.java
@@ -7,6 +7,8 @@ import ru.dpopkov.knowthenics.model.Answer;
 import ru.dpopkov.knowthenics.repositories.AnswerRepository;
 import ru.dpopkov.knowthenics.services.AnswerService;
 
+import java.util.Set;
+
 @Service
 @Profile("spring-data-jpa")
 public class AnswerSDJpaService extends AbstractSDJpaService<Answer> implements AnswerService {
@@ -22,5 +24,11 @@ public class AnswerSDJpaService extends AbstractSDJpaService<Answer> implements 
         } catch (NotFoundInRepositoryException ex) {
             throw new NotFoundInRepositoryException("Answer not found for ID " + id, ex);
         }
+    }
+
+    @Override
+    public Set<Answer> findAllByWordingEnLike(String searchString) {
+        AnswerRepository answerRepository = (AnswerRepository) super.crudRepository;
+        return answerRepository.findAllByWordingEnLike(searchString);
     }
 }