+++ knowthenics-web/src/main/java/ru/dpopkov/knowthenics/controllers/AnswersAllController.java
@@ -3,15 +3,22 @@ package ru.dpopkov.knowthenics.controllers;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import ru.dpopkov.knowthenics.model.Answer;
 import ru.dpopkov.knowthenics.services.AnswerService;
+
+import java.util.ArrayList;
+import java.util.Set;
 
 @SuppressWarnings("SameReturnValue")
 @Slf4j
 @Controller
 public class AnswersAllController {
+
+    private static final String ANSWERS_FIND_ANSWERS = "answers/find-answers";
 
     private final AnswerService answerService;
 
@@ -35,8 +42,27 @@ public class AnswersAllController {
     }
 
     @GetMapping("/answers/find")
-    public String find() {
-        log.debug("Finding Answers is not implemented yet");
-        return "notimplemented";
+    public String initFindAnswerForm(Model model) {
+        model.addAttribute(new Answer());
+        return ANSWERS_FIND_ANSWERS;
+    }
+
+    @PostMapping("/answers/find")
+    public String processFindAnswerForm(Answer answer, BindingResult result, Model model) {
+        if (answer.getWordingEn() == null) {
+            answer.setWordingEn("");
+        }
+        String searchPattern = "%" + answer.getWordingEn() + "%";
+        Set<Answer> found = answerService.findAllByWordingEnLike(searchPattern);
+        if (found.isEmpty()) {
+            result.rejectValue("wordingEn", "notFound", "not found");
+            return ANSWERS_FIND_ANSWERS;
+        } else if (found.size() == 1) {
+            Answer single = new ArrayList<>(found).get(0);
+            return "redirect:/questions/" + single.getQuestion().getId() + "/answers/" + single.getId() + "/view";
+        } else {
+            model.addAttribute("answers", found);
+            return "answers/list";
+        }
     }
 }